Move exception utils to error
