+++ app/src/main/java/org/dicio/dicio_android/error/ExceptionUtils.java
@@ -1,4 +1,4 @@
-package org.dicio.dicio_android.util;
+package org.dicio.dicio_android.error;
 
 import androidx.annotation.Nullable;
